remote: capture tags at message load
Interpreting JMAP keywords is not really something the local store needs
to deal with; we can just bundle the tags directly in the remote mail
object.

Moving the add-mailboxes-to-tags work up into sync might be a temporary
lift; I think I want to move it into remote proper but that needs a
little more work. Either way though, I'm fairly sure it does not belong
in local::Email.