Specify full mailbox list when pushing
Instead of using JSON pointer/patch syntax for specifying mailboxIds,
instead explicitly list the full set of mailboxes for a message. This is
an attempt to fix #19.

This also fixes an edge-case where if a message should belong to an
ignored mailbox and no others, mujmap would have mistakenly added it to
the archive mailbox. This wouldn't strictly cause issues, but it's not
desired behavior nevertheless.