local: get mail dir from notmuch config
notmuch has two path config variables, `database.path` and
`database.mail_root`. Setting them both allows the db and mail dirs to
be in separate parts of the filesystem, which is typically needed to use
NOTMUCH_PROFILE and/or XDG config/storage dirs.

Not setting `mail_root` makes notmuch use `path` as the root, with the
real database in `.notmuch`, as normal.

To allow this to work with mujmap, this commit removes the restriction
that the mail dir is under the db dir, and always uses notmuch's idea of
where the mail dir should be.