local: compute all-mail query for external paths
notmuch has an idea of where the mail dir is, which is the "root" for
its path queries. If mujmap's mail dir is under mail root, then compute a
path query relative to it. Otherwise, if mujmap's mail dir is the same
as the mail root, or somewhere outside it entirely, use a "global" path
query.