config: ensure configured `mail_dir` exists
This is a small protection against typos in this config item. Its
possible that you actually wanted a subdir of the notmuch mail_root, but
typed it wrong. That would lead notmuch to miscompute the notmuch path
prefix, which could result in mujmap having the wrong idea about what
mail exists locally or on the server, with a risk of mail being deleted
from one or both ends.

Its hard to say what the odds are of a mistyped path matching a valid
directory, but it seems reasonably unlikely, so hopefully this will
prevent the most common case.