remote: reorganize initial server authentication
Previously, mujmap would always blindly do HTTP Basic authentication
without doing an initial request to see if the server actually wants
authentication, or what kind of authentication it wants. That's not
wrong, but isn't enough. We will soon see servers supporting Bearer
token authentication, so we need to first ask the server if it requires
authentication before we proceed.

This commit reorganises the auth and session setup to do that. Now, when
retrieving the initial session we do an unauthenticate get. If that
fails, we then go on to set up for Basic authentication, and make sure
that succeeds before proceeding.

This has the somewhat strange side-effect that the server might actually
allow the initial session to be fetched unauthenticated. I would be
surprised to see this in the wild, but its not actually wrong and I can
imagine a situation where its possible for internal servers.

This restructure gives an obvious hook point to add alternative
authentication schemes, like Bearer. See the next commit for that!