remote: ensure configured username matches authenticated session username
Bearer tokens don't require a username to authenticate, so its possible
to use the "wrong" token and everything end up seem fine. In that case,
config.username would be wrong and though right now that wouldn't affect
anything, it could in the future and even if not, still means the user
is perhaps in a muddle. Best to notice it early and bail out, since its
almost certainly an easy fix in the config.