config: trim whitespace from password
The password command will often emit a trailing newline, which mujmap
was including in the password sent on the wire. Basic auth users base64,
which is sort of lenient about newlines, but other auth schemes are not
quite as forgiving.

Removing all trailing and leading whitespace is possibly too aggressive,
as it will break passwords that have legitimate whitespace in those
positions. On the other hand, its easy to do this way and such a
password seems bananas, so I'll take the risk for now.